Refactor Rakefile into tasks, refactor task execution
The Rakefile tasks are now broken up into smaller .rake files (i.e.,
one for vim, symlink, ruby, brew, zsh, etc...). In addition, I made it
so that the task simply performed the actions instead of delegating to a
method of a similar name.

There is also a utils.rb which contains common methods across all .rake
files.

Everything with the ./tasks dir goes through a 'load' in the Rakefile.

The brew (cask) packages have a slight update to them.