VIM: Update most configs to use `vim` or `nvim` configs based on env
As `vim` and `nvim` are two different executables they both need their
own set of plugins and configurations. This commit allows `vim` and
`nvim` too be separate and vary on global variables and conditionals.

Update the vim.rake task to install vundle and update vim for both
`vim` and `nvim`.