Vim: Replace yankring with vim-easyclip
Yankring just didn't work with neovim, so thats out.

vim-easyclip has most of Yankring's functionality (cycle through the
yank buffer). It also introduces the 's' and 'm' actions to substitute
and cut. These actions remove the issue with deleting and changing
modifying what is in the registers. We will see how this works out,
it'll take a little getting use to initially.