Use Mackup to backup certain applications to dotfiles repo
A two config approach is used. One config will be for the dotfiles
applications (i.e., zsh, neovim, bin, etc...). The other will be for all
the other configurations I don't care too much on sharing (i.e.,
applications), which will be synced to Dropbox.

The plumbing to deal with the two configs will be handled later via rake
tasks.

Update .gitignore to account for the dynamic/large/generated files that
might be included in the Mackup directory.

Include mackup,  neovim, zsh, tips, ruby, bin and git in mackup backup.