Setup the two mackup configs for local and dropbox, include mackup tasks
We have a local and dropbox configuration for mackup. These configs in
combination with the mackup task allows to backup 'sharable' configs
locally to the dotfiles repo, and the 'unimportant' configs in dropbox.

This approach isn't ideal, as we're juggling configs, but it will keep a
separation in the applications that are synced to each destination.

A default mackup config is created to be a placeholder (which syncs
nothing except the mackup configs themselves).

Idea is to interact to Mackup through the rake tasks. Although knowing
what you are doing you should be fine with manual control.

In the future maybe there will be a better way to handle this.