Fix .gitignore to actually whitelist recursively in directories
The change here makes it so all files are originally blacklisted, and
we now have all directories whitelisted. This means that in the
whitelisting section we can use recursive wildcards to whitelist under
that directory. A second blacklisting is needed to filter out the stuff
from the recursive whitelisting.

Add in missing files due to this issue.
 - bin/blog (for opening quickly to blog)
 - .mackup/ssh.cfg
 - tasks/mas.rake