Mass cleanup of brew casks/packages/taps and Mac App Store backups
A bunch of stuff has accumulated over the years... This is an attempt to
drop things I don't use anymore. In addition, it should reduce the
choice of tools and applications.

I'll likely slowly re-add some of these over time, but at least I have
a smaller surface area to work off of now.