feat(vim): full revamp of the nvim setup
The main thing is is pairing down on certain aspect I never used.

I also swapped CtrlP for telescope (although need to still get use to it
all), and youcompleteme for nvim-compe.

Switched the themes from tomorrow-night to gruvbox.

It is highly possible some things might get lost in this commit, but
overall things are 'smaller' as I use VScode more and more now.